requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile
#
asgiref==3.7.2
    # via
    #   channels
    #   channels-redis
    #   django
    #   django-cors-headers
async-timeout==4.0.3
    # via redis
backports-zoneinfo==0.2.1
    # via django
blinker==1.7.0
    # via flask
channels==4.0.0
    # via
    #   -r requirements.in
    #   channels-redis
    #   django-plotly-dash
channels-redis==4.1.0
    # via -r requirements.in
click==8.1.7
    # via flask
dash==2.9.3
    # via
    #   -r requirements.in
    #   dash-bootstrap-components
    #   django-plotly-dash
dash-bootstrap-components==1.5.0
    # via django-plotly-dash
dash-core-components==2.0.0
    # via dash
dash-html-components==2.0.0
    # via dash
dash-renderer==1.9.1
    # via -r requirements.in
dash-table==5.0.0
    # via dash
django==4.2.8
    # via
    #   -r requirements.in
    #   channels
    #   django-cors-headers
    #   django-plotly-dash
    #   djangorestframework
    #   djangorestframework-simplejwt
django-cors-headers==4.3.1
    # via -r requirements.in
django-plotly-dash==2.2.2
    # via -r requirements.in
djangorestframework==3.14.0
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==5.3.1
    # via -r requirements.in
dpd-components==0.1.0
    # via django-plotly-dash
dpd-static-support==0.0.5
    # via -r requirements.in
flask==3.0.0
    # via
    #   dash
    #   django-plotly-dash
imageio==2.33.1
    # via scikit-image
importlib-metadata==7.0.1
    # via flask
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via flask
lazy-loader==0.3
    # via scikit-image
markupsafe==2.1.3
    # via
    #   jinja2
    #   werkzeug
msgpack==1.0.7
    # via channels-redis
networkx==3.1
    # via scikit-image
numpy==1.24.4
    # via
    #   imageio
    #   pandas
    #   pywavelets
    #   scikit-image
    #   scipy
    #   tifffile
packaging==23.2
    # via
    #   plotly
    #   scikit-image
pandas==2.0.3
    # via -r requirements.in
pillow==10.1.0
    # via
    #   -r requirements.in
    #   imageio
    #   scikit-image
plotly==5.18.0
    # via
    #   -r requirements.in
    #   dash
    #   django-plotly-dash
pyjwt==2.8.0
    # via djangorestframework-simplejwt
python-dateutil==2.8.2
    # via pandas
pytz==2023.3.post1
    # via
    #   djangorestframework
    #   pandas
pywavelets==1.4.1
    # via scikit-image
redis==5.0.1
    # via channels-redis
scikit-image==0.21.0
    # via -r requirements.in
scipy==1.10.1
    # via scikit-image
six==1.16.0
    # via python-dateutil
sqlparse==0.4.4
    # via django
tenacity==8.2.3
    # via plotly
tifffile==2023.7.10
    # via scikit-image
typing-extensions==4.9.0
    # via asgiref
tzdata==2023.3
    # via pandas
werkzeug==3.0.1
    # via
    #   django-plotly-dash
    #   flask
whitenoise==6.6.0
    # via -r requirements.in
zipp==3.17.0
    # via importlib-metadata
